made a hook and fixed offensive fumbler logic
