+++ components/BoxScore/PlayerData.jsx
@@ -5,16 +5,13 @@
 //   "passing",
 //   "rushing",
 //   "receiving",
-//   "fumbles",
 //   "defensive",
-//   "interceptions",
 //   "kick_returns",
 //   "kicking",
 //   "punting",
 //   "punt_returns",
 // ];
 
-// // Define which stats to display and their labels for each group
 // const DISPLAY_STATS = {
 //   passing: {
 //     "comp att": "C/ATT",
@@ -57,19 +54,111 @@
 //     average: "AVG",
 //     "punt return td": "TD",
 //   },
+//   defensive: {
+//     tackles: "TOT",
+//     tfl: "TFL",
+//     sacks: "SACKS",
+//     fr: "FR", // custom-added
+//     int: "INT", // custom-added
+//     td: "TD", // custom-added
+//   },
 // };
 
 // export function PlayerData({ stats }) {
+//   // --- Step 1: Normalize group names ---
+//   const normalizeGroups = {};
+//   stats.forEach((group) => {
+//     const key = group.name.toLowerCase().replace(/\s/g, "_");
+//     normalizeGroups[key] = group;
+//   });
+
+//   // --- Step 2: Build merged defensive group ---
+//   const defensiveMap = new Map();
+
+//   const addOrUpdatePlayer = (player, fields) => {
+//     const id = player.player.id;
+//     if (!defensiveMap.has(id)) {
+//       defensiveMap.set(id, {
+//         id,
+//         name: player.player.name,
+//         tackles: 0,
+//         tfl: 0,
+//         sacks: 0,
+//         fr: 0,
+//         int: 0,
+//         td: 0,
+//       });
+//     }
+
+//     const entry = defensiveMap.get(id);
+//     for (const [key, value] of Object.entries(fields)) {
+//       entry[key] += value ? parseInt(value, 10) || 0 : 0;
+//     }
+//   };
+
+//   // Add from Defensive group
+//   if (normalizeGroups.defensive) {
+//     normalizeGroups.defensive.players.forEach((player) => {
+//       const fields = {};
+//       player.statistics.forEach((s) => {
+//         if (s.name === "tackles") fields.tackles = s.value;
+//         if (s.name === "tfl") fields.tfl = s.value;
+//         if (s.name === "sacks") fields.sacks = s.value;
+//         if (s.name === "interceptions for touch downs") fields.td = s.value;
+//       });
+//       addOrUpdatePlayer(player, fields);
+//     });
+//   }
+
+//   // Add from Fumbles group
+//   if (normalizeGroups.fumbles) {
+//     normalizeGroups.fumbles.players.forEach((player) => {
+//       const fields = {};
+//       player.statistics.forEach((s) => {
+//         if (s.name === "rec") fields.fr = s.value;
+//         if (s.name === "rec td") fields.td = s.value;
+//       });
+//       addOrUpdatePlayer(player, fields);
+//     });
+//   }
+
+//   // Add from Interceptions group
+//   if (normalizeGroups.interceptions) {
+//     normalizeGroups.interceptions.players.forEach((player) => {
+//       const fields = {};
+//       player.statistics.forEach((s) => {
+//         if (s.name === "total interceptions") fields.int = s.value;
+//       });
+//       addOrUpdatePlayer(player, fields);
+//     });
+//   }
+
+//   // --- Step 3: Override the defensive group with the merged one ---
+//   const mergedDefensiveGroup = {
+//     name: "Defensive",
+//     players: Array.from(defensiveMap.values()).map((player) => ({
+//       player: { id: player.id, name: player.name },
+//       statistics: [
+//         { name: "tackles", value: player.tackles },
+//         { name: "tfl", value: player.tfl },
+//         { name: "sacks", value: player.sacks },
+//         { name: "fr", value: player.fr },
+//         { name: "int", value: player.int },
+//         { name: "td", value: player.td },
+//       ],
+//     })),
+//   };
+
+//   normalizeGroups.defensive = mergedDefensiveGroup;
+
 //   const renderStatGroup = (groupName) => {
-//     const group = stats.find(
-//       (g) => g.name.toLowerCase().replace(/\s/g, "_") === groupName
-//     );
+//     const group = normalizeGroups[groupName];
 
 //     if (!group) {
 //       return <Txt key={groupName}>No stats for {groupName.replace(/_/g, " ")}</Txt>;
 //     }
 
-//     const statLabelMap = DISPLAY_STATS[groupName] ?? null;
+//     const statLabelMap = DISPLAY_STATS[groupName];
 
 //     return (
 //       <View key={groupName} style={s.groupContainer}>
@@ -93,11 +182,7 @@
 //     );
 //   };
 
-//   return (
-//     <View>
-//       {STAT_GROUPS.map((groupName) => renderStatGroup(groupName))}
-//     </View>
-//   );
+//   return <View>{STAT_GROUPS.map((groupName) => renderStatGroup(groupName))}</View>;
 // }
 
 // const s = StyleSheet.create({
@@ -127,6 +212,7 @@
 
 import { StyleSheet, View } from "react-native";
 import { Txt } from "../general/Txt";
+import { usePlayerStats } from "../../hooks/usePlayerStats"; // adjust import path
 
 const STAT_GROUPS = [
   "passing",
@@ -158,17 +244,8 @@ const DISPLAY_STATS = {
     average: "AVG",
     "receiving touch downs": "TD",
   },
-  kicking: {
-    "field goals": "FG",
-    long: "LONG",
-    "extra point": "XP",
-  },
-  punting: {
-    total: "NO",
-    yards: "YDS",
-    average: "AVG",
-    touchbacks: "TB",
-  },
+  kicking: { "field goals": "FG", long: "LONG", "extra point": "XP" },
+  punting: { total: "NO", yards: "YDS", average: "AVG", touchbacks: "TB" },
   kick_returns: {
     total: "NO",
     yards: "YDS",
@@ -185,109 +262,24 @@ const DISPLAY_STATS = {
     tackles: "TOT",
     tfl: "TFL",
     sacks: "SACKS",
-    fr: "FR", // custom-added
-    int: "INT", // custom-added
-    td: "TD", // custom-added
+    fr: "FR",
+    int: "INT",
+    td: "TD",
   },
 };
 
 export function PlayerData({ stats }) {
-  // --- Step 1: Normalize group names ---
-  const normalizeGroups = {};
-  stats.forEach((group) => {
-    const key = group.name.toLowerCase().replace(/\s/g, "_");
-    normalizeGroups[key] = group;
-  });
-
-  // --- Step 2: Build merged defensive group ---
-  const defensiveMap = new Map();
-
-  const addOrUpdatePlayer = (player, fields) => {
-    const id = player.player.id;
-    if (!defensiveMap.has(id)) {
-      defensiveMap.set(id, {
-        id,
-        name: player.player.name,
-        tackles: 0,
-        tfl: 0,
-        sacks: 0,
-        fr: 0,
-        int: 0,
-        td: 0,
-      });
-    }
-
-    const entry = defensiveMap.get(id);
-    for (const [key, value] of Object.entries(fields)) {
-      entry[key] += value ? parseInt(value, 10) || 0 : 0;
-    }
-  };
-
-  // Add from Defensive group
-  if (normalizeGroups.defensive) {
-    normalizeGroups.defensive.players.forEach((player) => {
-      const fields = {};
-      player.statistics.forEach((s) => {
-        if (s.name === "tackles") fields.tackles = s.value;
-        if (s.name === "tfl") fields.tfl = s.value;
-        if (s.name === "sacks") fields.sacks = s.value;
-        if (s.name === "interceptions for touch downs") fields.td = s.value;
-      });
-      addOrUpdatePlayer(player, fields);
-    });
-  }
-
-  // Add from Fumbles group
-  if (normalizeGroups.fumbles) {
-    normalizeGroups.fumbles.players.forEach((player) => {
-      const fields = {};
-      player.statistics.forEach((s) => {
-        if (s.name === "rec") fields.fr = s.value;
-        if (s.name === "rec td") fields.td = s.value;
-      });
-      addOrUpdatePlayer(player, fields);
-    });
-  }
-
-  // Add from Interceptions group
-  if (normalizeGroups.interceptions) {
-    normalizeGroups.interceptions.players.forEach((player) => {
-      const fields = {};
-      player.statistics.forEach((s) => {
-        if (s.name === "total interceptions") fields.int = s.value;
-        // if (s.name === "intercepted touch downs") fields.td = s.value;
-      });
-      addOrUpdatePlayer(player, fields);
-    });
-  }
-
-  // --- Step 3: Override the defensive group with the merged one ---
-  const mergedDefensiveGroup = {
-    name: "Defensive",
-    players: Array.from(defensiveMap.values()).map((player) => ({
-      player: { id: player.id, name: player.name },
-      statistics: [
-        { name: "tackles", value: player.tackles },
-        { name: "tfl", value: player.tfl },
-        { name: "sacks", value: player.sacks },
-        { name: "fr", value: player.fr },
-        { name: "int", value: player.int },
-        { name: "td", value: player.td },
-      ],
-    })),
-  };
-
-  normalizeGroups.defensive = mergedDefensiveGroup;
+  const normalizedGroups = usePlayerStats(stats);
 
   const renderStatGroup = (groupName) => {
-    const group = normalizeGroups[groupName];
-
+    const group = normalizedGroups[groupName];
     if (!group) {
-      return <Txt key={groupName}>No stats for {groupName.replace(/_/g, " ")}</Txt>;
+      return (
+        <Txt key={groupName}>No stats for {groupName.replace(/_/g, " ")}</Txt>
+      );
     }
 
     const statLabelMap = DISPLAY_STATS[groupName];
-
     return (
       <View key={groupName} style={s.groupContainer}>
         <Txt style={s.groupTitle}>{group.name}</Txt>
@@ -295,13 +287,10 @@ export function PlayerData({ stats }) {
           <View key={player.player.id} style={s.playerContainer}>
             <Txt style={s.playerName}>{player.player.name}</Txt>
             {player.statistics
-              .filter((stat) =>
-                statLabelMap ? Object.keys(statLabelMap).includes(stat.name) : true
-              )
+              .filter((stat) => statLabelMap?.[stat.name])
               .map((stat, index) => (
                 <Txt key={index} style={s.statLine}>
-                  {(statLabelMap && statLabelMap[stat.name]) || stat.name}:{" "}
-                  {stat.value ?? "N/A"}
+                  {statLabelMap[stat.name]}: {stat.value ?? "N/A"}
                 </Txt>
               ))}
           </View>
@@ -310,30 +299,28 @@ export function PlayerData({ stats }) {
     );
   };
 
-  return <View>{STAT_GROUPS.map((groupName) => renderStatGroup(groupName))}</View>;
+  return <View>{STAT_GROUPS.map(renderStatGroup)}</View>;
 }
 
 const s = StyleSheet.create({
-  groupContainer: {
-    marginBottom: 16,
-  },
+  groupContainer: { marginBottom: 16 },
   groupTitle: {
     fontSize: 16,
     fontWeight: "bold",
     marginBottom: 4,
     color: "#54D18C",
   },
-  playerContainer: {
-    marginLeft: 12,
-    marginBottom: 8,
+  playerContainer: { 
+    marginLeft: 12, 
+    marginBottom: 8 
   },
-  playerName: {
-    fontWeight: "600",
-    marginBottom: 2,
-    color: "#F8F8F8",
+  playerName: { 
+    fontWeight: "600", 
+    marginBottom: 2, 
+    color: "#F8F8F8" 
   },
-  statLine: {
-    marginLeft: 8,
-    color: "#B8C3CC",
+  statLine: { 
+    marginLeft: 8, 
+    color: "#B8C3CC" 
   },
 });
